requirements.txt
streamlit>=1.28.0
pdfplumber>=0.10.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
openai>=1.0.0

